requirements.txt
openai>=1,<2
flask>=3
pillow
reportlab
python-dotenv
requests
